Added Gameplay Manager
The gameplay manager will handle stuff like pausing, probably spawning characters, etc. It is not done, and has a few TODO statements for the character/AI team to look at.

I temporarily added Character and ActiveObject scripts. If these are changed to something, then so be it. I wanted to appease Unity to get rid of compile errors.